+++ CHANGES.md
@@ -1,5 +1,6 @@
 # ?.?.?
 - add: add quick filter buttons for track types on track and month goal page
+- add: add quick filter buttons for track types and years on map page
 - add: new chart "speed per track"
 - fix: broken search page and improve responsiveness
 - fix: custom fields: show display name for type instead of enum name on settings page